+++ src/com/ansorgit/plugins/bash/lang/psi/impl/command/BashFunctionVariantsProcessor.java
@@ -19,6 +19,7 @@ import com.ansorgit.plugins.bash.lang.psi.api.function.BashFunctionDef;
 import com.ansorgit.plugins.bash.lang.psi.util.BashAbstractProcessor;
 import com.ansorgit.plugins.bash.lang.psi.util.BashPsiUtils;
 import com.google.common.collect.Lists;
+import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.util.Key;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.ResolveState;
@@ -38,6 +39,8 @@ public class BashFunctionVariantsProcessor extends BashAbstractProcessor {
     }
 
     public boolean execute(PsiElement element, ResolveState resolveState) {
+        ProgressManager.checkCanceled();
+
         if (element instanceof BashFunctionDef) {
             final BashFunctionDef f = (BashFunctionDef) element;