+++ src/com/ansorgit/plugins/bash/lang/psi/impl/vars/DumbVarDefReference.java
@@ -34,7 +34,7 @@ class DumbVarDefReference extends AbstractVarDefReference {
             return null;
         }
 
-        return BashResolveUtil.resolve(bashVarDef, true, true);
+        return BashResolveUtil.resolve(bashVarDef, true, false);
 
     }
 }