+++ Changelog.md
@@ -4,6 +4,8 @@
  - [#753](https://github.com/BashSupport/BashSupport/issues/753): Fix NPE in BashRunConfiguration.suggestedName()
  - [#758](https://github.com/BashSupport/BashSupport/issues/758): Completions often showing "Invalid" as label
  - [#761](https://github.com/BashSupport/BashSupport/issues/761): IOException while scanning path
+ - [#764](https://github.com/BashSupport/BashSupport/issues/764): Variables declared in included files were not always properly resolved.
+ - Fix exception about IndexPatternBuilder incorrect Bash comment ranges
 
 #### 2019-07-04:
  - A notification is displayed for 192.x and later if both BashSupport and JetBrains Shell are enabled at the same time. It allows to disable one of the plugins.